requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关已移除

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 环境变量管理
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP 客户端
httpx==0.25.2

# 日志
loguru==0.7.2

# 工具库
python-dateutil==2.8.2

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8